move relevant things around
